requirements.txt
djangorestframework==3.14.0
django==5.0.1
djoser==2.2.2
django-cors-headers==4.3.1
django_db_logger==0.1.13
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
python-json-logger==2.0.7
python-docx=1.1.2